removed express-helper and migrated its functionality within express
